server: refactor sudoku functions into separate file
